+++ src/nsidc/iceflow/data/atm1b.py
@@ -116,36 +116,38 @@ def _file_dtype(filepath: Path) -> DTypeLike:
     return dtype
 
 
-def _blatm1bv1_date(fn):
+def _blatm1bv1_date(fn) -> dt.date:
     """Return the date from the given BLATM1B filename."""
     fn_date = None
 
     m = re.search(r"BLATM1B_(\d{8})", fn)
     if m:
-        fn_date = m.group(1)
+        fn_date_str = m.group(1)
     else:
         m = re.search(r"BLATM1B_(\d{6})", fn)
-        if m:
-            fn_date = m.group(1)
-            if int(fn_date[:2]) > 9:
-                fn_date = "19" + fn_date
-            else:
-                fn_date = "20" + fn_date
+        if not m:
+            err_msg = f"Failed to extract date from BLATM1B v1 file: {fn}"
+            raise RuntimeError(err_msg)
 
-    if fn_date:
-        fn_date = dt.datetime.strptime(fn_date, "%Y%m%d")
+        fn_date_str = m.group(1)
+        if int(fn_date_str[:2]) > 9:
+            fn_date_str = "19" + fn_date_str
+        else:
+            fn_date_str = "20" + fn_date_str
+
+    fn_date = dt.datetime.strptime(fn_date_str, "%Y%m%d").date()
 
     return fn_date
 
 
-def _ilatm1b_date(fn: str) -> dt.datetime:
+def _ilatm1b_date(fn: str) -> dt.date:
     """Return the date from the given ILATM1B filename."""
     m = re.search(r"_(\d{8})_", fn)
     if not m:
         err = f"Failed to extract date from filepath: {fn}"
         raise RuntimeError(err)
     fn_date = m.group(1)
-    return dt.datetime.strptime(fn_date, "%Y%m%d")
+    return dt.datetime.strptime(fn_date, "%Y%m%d").date()
 
 
 def _shift_lon(lon):
@@ -209,7 +211,7 @@ def _utc_datetime(
     tdf["month"] = file_date.month
     tdf["day"] = file_date.day
 
-    ls = int(leap_seconds(file_date))
+    ls = int(leap_seconds(dt.datetime(file_date.year, file_date.month, file_date.day)))
     utc = pd.to_datetime(tdf).to_numpy() - np.timedelta64(ls, "s")
 
     return pd.Series(utc)
@@ -260,16 +262,40 @@ def _atm1b_qfit_data(filepath: Path, file_date: dt.date) -> pd.DataFrame:
     return df
 
 
+# TODO: extract this for reuse with other data products.
+def _normalize_itrf_str(itrf_str: str) -> str:
+    """Normalizes common ITRF strings into ones recognizable by proj.
+
+    E.g., "ITRF2020" is common, but proj only recognizes "ITRF20".
+    """
+    itrf_str = itrf_str.upper()
+    try:
+        itrf_str = {
+            "ITRF00": "ITRF2000",
+            "ITRF05": "ITRF2005",
+            "ITRF08": "ITRF2008",
+            "ITRF2020": "ITRF20",
+        }[itrf_str]
+    except KeyError:
+        pass
+
+    return itrf_str
+
+
 def _qfit_file_header(filepath: Path) -> str:
     """Return the header string from a QFIT file."""
-    dtype = np.dtype([("record_type", ">i4"), ("header", ">S44")])
-
     record_size = np.fromfile(filepath, dtype=">i4", count=1)[0]
+    # The header length for each record is the number of bytes after the first
+    # word.
+    header_size = record_size - 4
+    dtype = np.dtype([("record_type", ">i4"), ("header", f">S{header_size}")])
+
     if record_size >= 100:
         record_size = np.fromfile(filepath, dtype="<i4", count=1)[0]
         if record_size >= 100:
             raise ValueError("invalid record size found")
-        dtype = np.dtype([("record_type", "<i4"), ("header", "<S44")])
+        header_size = record_size - 4
+        dtype = np.dtype([("record_type", "<i4"), ("header", f"<S{header_size}")])
 
     raw_data = np.fromfile(filepath, dtype=dtype)
 
@@ -286,40 +312,56 @@ def _qfit_file_header(filepath: Path) -> str:
     raise RuntimeError(err)
 
 
+def _infer_qfit_itrf(filepath: Path) -> str:
+    """Takes an ILATM1B/BLATM1B qfit filepath and returns a string representing
+    the ITRF.
+
+    This function infers the ITRF based on the qfit file header, which is
+    described here:
+    https://nsidc.org/sites/nsidc.org/files/files/ReadMe_qfit.txt
+
+    The string we extract the ITRF from looks like this:
+
+        `./091109_aa_l12_cfm_itrf05_18may10_palm_roth_amu2`
+
+    From which we extract an ITRF of `ITRF2005` from the `_itrf05_` bit.
+
+    According to Michael Studinger (see
+    https://github.com/nsidc/iceflow/issues/35#issuecomment-2408619586), This
+    string represents the "GPS trajectory that was used to reference the lidar
+    data and that has the ITRF epoch in its file name."
+    """
+    header = _qfit_file_header(filepath)
+    results = re.finditer(r"itrf\d{2,4}", header)
+    itrfs = list({result.group() for result in results})
+
+    if len(itrfs) == 1:
+        itrf_from_qfit_header = _normalize_itrf_str(itrfs[0])
+        return itrf_from_qfit_header
+
+    err_msg = f"Failed to extract ITRF from qfit header: {filepath}"
+    raise RuntimeError(err_msg)
+
+
+def _extract_itrf_from_h5_file(filepath: Path) -> str:
+    """Extract ITRF from the h5 filepath (ILATM1B v2)."""
+    with h5py.File(filepath, "r") as ds:
+        itrf = str(ds["ancillary_data"]["reference_frame"][:][0], encoding="utf8")
+    itrf = _normalize_itrf_str(itrf)
+
+    return itrf
+
+
 def extract_itrf(filepath: Path) -> str:
     ext = filepath.suffix
 
     if ext == ".qi":
-        header = _qfit_file_header(filepath)
-        results = re.finditer(r"itrf\d{2,4}", header)
-        itrfs = list({result.group() for result in results})
+        return _infer_qfit_itrf(filepath)
+    elif ext == ".h5":
+        return _extract_itrf_from_h5_file(filepath)
 
-        if len(itrfs) == 1:
-            itrf = itrfs[0]
-        else:
-            err = f"Failed to extract ITRF from file: {filepath}"
-            raise RuntimeError(err)
-
-    elif ext == "h5py":
-        with h5py.File(filepath, "r") as ds:
-            itrf = str(ds["ancillary_data"]["reference_frame"][:][0], encoding="utf8")
-    else:
-        err = f"Failed to read ITRF from unrecognized file: {filepath}"
-        raise RuntimeError(err)
-
-    itrf = itrf.upper()
-
-    # Try to normalize based on known ITRF strings in the ATM1B data:
-    # TODO: extract this for reuse with other data products.
-    try:
-        itrf = {
-            "ITRF05": "ITRF2005",
-            "ITRF08": "ITRF2008",
-        }[itrf]
-    except KeyError:
-        pass
-
-    return itrf
+    err = f"Failed to read ITRF from unrecognized file: {filepath}"
+    raise RuntimeError(err)
 
 
 def _ilatm1bv2_dataframe(filepath: Path) -> pd.DataFrame:
@@ -384,10 +426,11 @@ def atm1b_data(filepath: Path) -> ATM1BDataFrame:
     # ILATM1B_20140430_110310.ATM4BT4.h5
     # ILATM1B_20111104_181304.ATM4BT4.qi
     # BLATM1B_20060522_145449.qi
+    # BLATM1B_20041127atm2_210316jr.lutF.qi
     filename = filepath.name
 
     # Find the date, which corresponds to the product version.
-    match = re.search(r".*_(\d{4})\d{4}_.*", filename)
+    match = re.search(r".*_(\d{4})\d{4}.*", filename)
     if not match:
         err = f"Failed to recognize {filename} as ATM1B data."
         raise RuntimeError(err)
@@ -395,11 +438,14 @@ def atm1b_data(filepath: Path) -> ATM1BDataFrame:
     year = int(match.group(1))
 
     if year >= 2013:
-        data = _ilatm1bv2_data(filepath, _ilatm1b_date(filename))
+        file_date = _ilatm1b_date(filename)
+        data = _ilatm1bv2_data(filepath, file_date)
     elif year >= 2009:
-        data = _atm1b_qfit_data(filepath, _ilatm1b_date(filename))
+        file_date = _ilatm1b_date(filename)
+        data = _atm1b_qfit_data(filepath, file_date)
     else:
-        data = _atm1b_qfit_data(filepath, _blatm1bv1_date(filename))
+        file_date = _blatm1bv1_date(filename)
+        data = _atm1b_qfit_data(filepath, file_date)
 
     itrf = extract_itrf(filepath)
     data["ITRF"] = itrf