+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   `DatasetSearchParameters`.
 - `transform_itrf` will calculate plate for source ITRF if not given with
   `target_epoch`.
+- Add support for ILATM1B v2 and BLATM1B v1.
 
 # v0.2.0